# src/controllers/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DOM Cloud",
    "version": "1.0.0",
    "description": "DOM Cloud Bridge API documentation"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": ""
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/nftables/show": {
      "get": {
        "summary": "Show nftables rules",
        "tags": [
          "Nftables"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[\\w.-]+$"
            },
            "description": "Filter rules by username. Must be alphanumeric with dots or dashes."
          }
        ],
        "responses": {
          "200": {
            "description": "Returns full or user-specific nftables rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/nftables/add": {
      "post": {
        "summary": "Add user to nftables rule",
        "tags": [
          "Nftables"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "user"
                ],
                "properties": {
                  "user": {
                    "type": "string",
                    "pattern": "^[\\w.-]+$",
                    "description": "Username to allow (alphanumeric, dot, dash only)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result of adding user to nftables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/nftables/del": {
      "post": {
        "summary": "Remove user from nftables rule",
        "tags": [
          "Nftables"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "user"
                ],
                "properties": {
                  "user": {
                    "type": "string",
                    "pattern": "^[\\w.-]+$",
                    "description": "Username to remove (alphanumeric, dot, dash only)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result of removing user from nftables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/status/check": {
      "get": {
        "summary": "Check daemon statuses",
        "tags": [
          "Status"
        ],
        "responses": {
          "200": {
            "description": "Normal daemon status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Abnormal daemon status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/status/test": {
      "get": {
        "summary": "Test server config for any wrong config",
        "tags": [
          "Status"
        ],
        "responses": {
          "200": {
            "description": "Normal server config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Abnormal server config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}